requirements.txt
gradio
numpy
pandas
transformers
faiss-cpu